requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.4.2
pydantic[email]

# Google API dependencies
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0

# Slack dependencies
slack-sdk==3.26.2
requests==2.31.0

# Additional utilities
httpx==0.25.0
python-multipart==0.0.6  # For file uploads

# Development dependencies (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0